% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{pipe}
\alias{pipe}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using magrittr semantics.}
}
\value{
The result of applying \code{rhs} to \code{lhs}.
}
\description{
See \code{\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
