% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_funs.R
\name{py_syndat_scores}
\alias{py_syndat_scores}
\title{Compute utility scores using the \code{syndat} Python module}
\usage{
py_syndat_scores(ori, syn, control = NULL)
}
\arguments{
\item{ori}{A \code{data.frame} or matrix with the original data used for
training.}

\item{syn}{A \code{data.frame} or matrix with the generated synthetic data.}

\item{control}{A \code{data.frame} providing hold-out control data against which
the synthetic data are compared.}
}
\value{
A named list with distribution, discrimination and correlation
scores for both the original and the control data.
}
\description{
The function bridges to the \code{syndat} Python package to evaluate different
quality metrics between an original and a synthetic data set.
}
