% Generated by hand; sync with R/utils-pipe.R
\docType{data}
\name{%>%}
\alias{%>%}
\title{Pipe operator}
\usage{lhs \%>\% rhs}
\arguments{
  \item{lhs}{A value or the magrittr placeholder.}
  \item{rhs}{A function call using the magrittr semantics.}
}
\description{
Forward a value into the next call. See \code{magrittr::\link[magrittr:pipe]{\%>\%}}
for full details.
}
\keyword{internal}
